Move Find sync code into a namespace
